fix: debug mode - encrypted api properties are not interpreted
